fix delete image from storage

+++ app/Http/Controllers/UserController.php
@@ -38,7 +38,9 @@ class UserController extends Controller
         if(request()->has('image')){
             $image_path = request()->file('image')->store('profile', 'public');
             $validated['image'] =$image_path;
-            Storage::disk('public')->delete($user->image);
+            if($user->image) {
+                Storage::disk('public')->delete($user->image);
+            }
         }
         $user->update($validated);
         return redirect()->route('profile');
